feat(java-shell): Cursor.next() should return JS object instead of Java object (#1065)
* feat(JMongosh): support conversion from Java to JS for all object types
* feat(JMongosh): cursor.next() should return JS object instead of Java object

+++ packages/java-shell/src/main/kotlin/com/mongodb/mongosh/result/MongoShellResult.kt
@@ -111,11 +111,11 @@ class DeleteResult(val acknowledged: Boolean, val deletedCount: Long) : MongoShe
 }
 
 class BulkWriteResult(val acknowledged: Boolean,
-                      val insertedCount: Long,
-                      val matchedCount: Long,
-                      val modifiedCount: Long,
-                      val deletedCount: Long,
-                      val upsertedCount: Long,
+                      val insertedCount: Int,
+                      val matchedCount: Int,
+                      val modifiedCount: Int,
+                      val deletedCount: Int,
+                      val upsertedCount: Int,
                       val upsertedIds: List<Any?>) : MongoShellResult<Map<String, Any>>() {
     override val value: Map<String, Any>
         get() = mapOf("acknowledged" to acknowledged,
